requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
psycopg2-binary==2.9.9
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
tqdm==4.66.4
    # via -r requirements.in
